2020 Transport Calibration
all demand side calibration is done for 2020. the overall number was more or less achieved by altering the income elasticity. Tech share weights was used to change the energy by a fuel in a particular subsector. Before doing that the overall number of each subsector should be matched using the subsector share weight file.