Added trn_pass_road as a supply sector and LDV and Bus as subsectors
Added trn_pass_road as a supply sector and LDV and Bus as subsectors to play alter sharewts.
Reason behind this activity is that 4W ownership in 2070 in GJ was turning out out be ~650, which makes no sense therefore, an S-shaped trajectory needs to be followed. which in implementation means increasing Bus sharewt or Decreasing LDV sharewt