Truck efficiency and cost correction
The efficiency of electric trucks was changed to start from 2X more efficient in 2025, to 2.5X more efficient in 2070. The cost was also changed starting with 3X more expensive in 2020 and reducing over time. Efficiency before 2025 was set back to default value. Logit of bus was also set back to default value.