Industry efficiency
Industry efficiency calculations were revised a little. High efficiency (ME) scenario for NZ and ME scenario for BAU. Two new xml were created that can be commented in or out based on the scenario. in HE and ME the efficiency value remains kind of similar, only in HE it reaches that value in 2050 while in ME it reaches that value in 2065.